update UnmarshalJSON to also attempt to unmarshal into struct
You were very high when you wrote this so it's garbage, but basically
what happens is that if you are using swagger code generation, it will
create its own structs and ignore our marshal function. This results in
value being marshalled as:
{
    "Value": value,
    "Valid": bool
}

But our marshaller marshals into just the base type.

So if Swagger does that, but we're using a Swagger generated client
with our own unmarshalling on the host side, we need to be able to
handle that they generated a marshaler that does it differently.

Again, you're high, but trust me, this is the right thing.